+++ graph_attention_student/experiments/vgd_single__megan2__aqsoldb.py
@@ -1,17 +1,23 @@
 import os
 import pathlib
 import typing as t
+import random
 
 import tensorflow as tf
 import tensorflow.keras as ks
 from pycomex.functional.experiment import Experiment
 from pycomex.utils import folder_path, file_namespace
 
+PATH: str = pathlib.Path(__file__).parent.absolute()
+
 # == DATASET PARAMETERS ==
-VISUAL_GRAPH_DATASET_PATH: str = os.path.expanduser('/media/ssd/.visual_graph_datasets/datasets/aqsoldb')
+
+VISUAL_GRAPH_DATASET_PATH: str = 'aqsoldb'
+TEST_INDICES_PATH = os.path.join(PATH, 'assets', 'test_indices__logp.json')
 USE_DATASET_SPLIT: t.Optional[int] = 0
 TRAIN_RATIO: float = 0.8
 NUM_EXAMPLES: int = 100
+NUM_TEST: int = 1000
 NUM_TARGETS: int = 1
 
 NODE_IMPORTANCES_KEY: t.Optional[str] = None
@@ -36,8 +42,9 @@ FIDELITY_FUNCS = [
     lambda org, mod: tf.nn.relu(mod - org),
     lambda org, mod: tf.nn.relu(org - mod),
 ]
-SPARSITY_FACTOR = 1.0
-CONCAT_HEADS = False
+SPARSITY_FACTOR: float = 1.0
+CONCAT_HEADS: bool = False
+FINAL_DROPOUT: float = 0.0
 
 # == TRAINING PARAMETERS ==
 BATCH_SIZE = 32
@@ -59,6 +66,31 @@ experiment = Experiment.extend(
     namespace=file_namespace(__file__),
     glob=globals()
 )
+
+@experiment.hook('filter_indices')
+def training_bootstrap(e: Experiment,
+                       train_indices: t.List[int],
+                       test_indices: t.List[int],
+                       rep: int,
+                       **kwargs,
+                       ):
+    """
+    This hook is being called right after the train / test split is created in the main experiment. The function 
+    receives the lists of integer train and test indices and is supposed to return a tuple (train_indices, test_indices).
+    
+    This specific hook implementation realizes training "bootstrapping" - in this process the training indices are 
+    randomly sampled from themselves such that randomly some are removed and some appear multiple times. The purpose 
+    of this is to create slightly different training distributions for the training of a model ensemble for example.
+    """
+
+    e.log('applying bootstrapping by subsampling the training indices...')
+
+    # random.choices() will randomly sample from the given list but is "drawing with replacement" which means 
+    # that elements are no longer unique and can appear multiple times. This also means that it is possible 
+    # that some elements do not appear at all.
+    train_indices = random.choices(train_indices, k=len(train_indices))
+
+    return train_indices, test_indices
 
 experiment.run_if_main()